Fix crash due to queue resurrection
When we're calling dispatch_get_current_queue(), ARC automatically
retains the result of the function and increments the external reference
count of the resulting dispatch queue.

The problem is dispatch queues assume that once their external reference
count goes to zero that the queue is no longer in use. By
re-incrementing, we trigger an assert within libdispatch.

This fix applies a bandaid to this issue. If we don't store the result
of the function in an ARC pointer, then we're not going to inadvetently
retain a dead queue. Instead, we'll use pointer comparisons to check for
equality:
- Since we're not going to use the queue for anything or persist it,
  there's no risk we accidentally extend the queue's lifetime in bad
  ways.
- If the queue is about to expire, we're still allowed to continue our
  block's work as long as we want.

I think it'd be great if we found an alternative to
dispatch_get_current_queue() that allows us to achieve the same
performance. I'll continue to think about this. In the meantime, this
appears to fix this crash.