Call a Future's observers outside of the Future's state lock
Avoids a reentrancy deadlock when a promise's 'result' is used from
within an observer callback. This is a behavior change only in the
following scenario:

1. A promise is sealed and starts running its observers
2. While fulfilling observers, a new observer is added (on the same
   thread or from on another thread)

Previously, the future would wait until all existing observers have
been run, then run the new observer. Now, the new observer will run
immediately, since the only thing being checked is that the promise
has been sealed. We could do more work to preserve the old behavior,
but it's better to not depend on such intricacies in the first place.